+++ src/lib/libast/stdio/vasprintf.c
@@ -27,7 +27,7 @@ vasprintf(char** s, const char* fmt, va_list args)
 
 	if (f = sfstropen())
 	{
-		v = sfvprintf(f, fmt, args);
+		v = (int)sfvprintf(f, fmt, args);
 		if (!(*s = strdup(sfstruse(f))))
 			v = -1;
 		sfstrclose(f);
